feat(sutstat): request args, response status range
Update sutstat to allow the caller to set additional arguments to the
GET request sent to see if the system is up (or ready).
Through the command line, users can now specify additional request
arguments with the `--request-arg KEY=VALUE` flag. Arguments will be
collected in a dictionary and passed into the call to `requests.get()`.

This example disables verification of certificates:
```bash
sutstat --url https://12.34.45.78 --request-arg verify=False
```

Eventually, that translates into a call similar to:
```python
requests.get("https://12.34.56.78", verify=False)
```
where the `False` from the command line ends up being a python boolean
False value. To pass a string, try something like:
```python
--request-arg 'key="value"'
```
or
```python
--request-arg "key=\"value\""
```

Also including two new flags to tune the values of the response status
code that are considered successful. The flag `--response-status-min`
allows the user to set the minimum status code that is accepted as
success. The flag `--response-status-max` allows the user to set the
maximum status code that is accepted as successful. Together, they
allow the user to set an inclusive range of status codes that are
considered successful.

For example, to accept status codes from 200-299 as successful:
```bash
sutstat --response-status-min 200 --response-status-max 299
```
The default is to accept any status code from 100-599 as successful.